chatbot_project/requirements.txt
alembic==1.12.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2023.7.22
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.10
django-datatables-view==1.20.0
django-elasticsearch-dsl==8.0
django-mysql==4.12.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
elastic-transport==8.10.0
elasticsearch==8.10.1
elasticsearch-dsl==8.9.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.7
mysqlclient==2.2.0
nltk==3.8.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
plotly==5.18.0
psycopg2==2.9.9
px==0.1.0
PyJWT==1.7.1
pymongo==4.5.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
simplejwt==2.0.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.2.4
stripe==7.4.0
tenacity==8.2.3
threadpoolctl==3.4.0
timedelta==2020.12.3
tqdm==4.66.2
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
utils==1.0.1
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
